Moved Constraint library into the tool.
