+++ lib/CMakeLists.txt
@@ -22,4 +22,3 @@ if(CLANG_ENABLE_STATIC_ANALYZER)
   add_subdirectory(StaticAnalyzer)
 endif()
 add_subdirectory(Format)
-add_subdirectory(Constraints)
