--- lib/Constraints/CMakeLists.txt
@@ -1,5 +1,0 @@
-set(LLVM_LINK_COMPONENTS support)
-
-add_clang_library(Constraints
-  Constraints.cpp
-)
