+++ tools/checked-c-convert/NewTyp.h
@@ -6,7 +6,7 @@
 #include "clang/Frontend/FrontendAction.h"
 #include "clang/Tooling/Tooling.h"
 
-#include "clang/Constraints/Constraints.h"
+#include "Constraints.h"
 
 #include "utils.h"
